change cast of thread_id for fixing build error and warnings

+++ spd_db_conn.cc
@@ -744,10 +744,10 @@ int spider_db_errorno(
           struct tm lt;
           struct tm *l_time = localtime_r(&cur_time, &lt);
           fprintf(stderr, "%04d%02d%02d %02d:%02d:%02d [WARN SPIDER RESULT] "
-            "to %ld: %d %s\n",
+            "to %lld: %d %s\n",
             l_time->tm_year + 1900, l_time->tm_mon + 1, l_time->tm_mday,
             l_time->tm_hour, l_time->tm_min, l_time->tm_sec,
-            (ulong) current_thd->thread_id, error_num,
+            (long long int) current_thd->thread_id, error_num,
             conn->db_conn->get_error());
         }
         if (!conn->mta_conn_mutex_unlock_later)
@@ -765,10 +765,10 @@ int spider_db_errorno(
         struct tm lt;
         struct tm *l_time = localtime_r(&cur_time, &lt);
         fprintf(stderr, "%04d%02d%02d %02d:%02d:%02d [ERROR SPIDER RESULT] "
-          "to %ld: %d %s\n",
+          "to %lld: %d %s\n",
           l_time->tm_year + 1900, l_time->tm_mon + 1, l_time->tm_mday,
           l_time->tm_hour, l_time->tm_min, l_time->tm_sec,
-          (ulong) current_thd->thread_id, error_num,
+          (long long int) current_thd->thread_id, error_num,
           conn->db_conn->get_error());
       }
       if (!conn->mta_conn_mutex_unlock_later)
